thinning out Node
